Added implicit management of SetMusicStyle before Similarity
These bindings now try to hide the musly jukebox and the
required setup process from the user. This means we add
and remove tracks from the jukebox as we see fit and call
SetMusicStyle on first Similarity usage.

Also found a single edge-case in libmusly code that allows it
to return a 2 on certain error conditions; We now handle this