cmd: add signal handling (again)
The setup is that the context passed to the command is canceled when a
signal is received to exit; We then wait for the command to return
before exiting from main.

This means any command that does not listen to context cancels is
also ignoring signals. Might be changed in later revisions to have a
force shutdown when multiple signals are send.