listener: added extra debug steps after a rare bug occured
The bug caused the listener to stop receiving any kind of metadata, but
the connection was still alive and receiving data. Due to the debugger
pausing the process and then resuming causing the connection to reset
the bug vanished before we found the source of the problem.

The added TODO is for the most likely cause, where we somehow got out of
sync with the interleaved metadata format and end up reading garbage
instead of actual metadata. Another way to counter this would be to parse
the actual mp3 data for validity.