irc: make AnnounceSong implementation actually use the argument
progress on issue #29

The announcement now actually uses the information passed to it as
argument, rather than asking for a new status from the configured
manager.

rpc: broke compatibility by changing field types to be easier to use;
functions in pkg time all use int64 for unixtime so we should be using
that too.

Bot.AnnounceSong now uses a separate message type to be able to
receive more required information.