queue: reworked QueueService interface
We've reworked the interface to be less of a traditional data structure
"queue" and more based on what we actually need from it as an audio
stream queue.

We also implement the new interface on the existing database-based queue
implementation that already existed. Some rough edges in terms of error
handling still exist

database: the Length field retrieved from the database is now a float64
to support more accurate times returned by the queue table

database: renamed UpdateTrackRequestTime to UpdateTrackRequestInfo to
better convey the fact it touches more than just the lastrequested field

streamer: updated to use the new queue interface