irc: implemented RequestTrack with friendly error messages
streamer: RequestSong now returns a more static and fleshed out
SongRequestError when the cause of failure is user controllable

rpc: added fields to RequestResponse to support the new fields in
SongRequestError. Helpers added to go between the two as well.

radio: implemented UntilRequestable that returns the amount of time left
before the song becomes requestable again