errors: added SelectDelay, to select a non-zero Delay field
This was required because the previous introduced Select returns the
*Error with the given kind, but because we pull up Kind's when
constructing an error in E we lose actual information associated with a
specific Kind.

SelectDelay changes this to just trying to find the first
non-zero Delay field such that the Kind doesn't matter anymore. It is
advised to call Is before calling SelectDelay to check that you have the
expected error kind.

rpc: fixed nil-pointer dereference in fromProtoError