migrations: added infrastructure for SQL migrations
This is a new directory containing .sql files that can migrate our
database schema up and down. We are going to try and do this with
https://github.com/golang-migrate/migrate and migrations files embedded
with https://github.com/tv42/becky such that a build of the project
always has the schema it requires with it.

The embedded files can be generated by running `go generate` in either
the root of the repository, or the migrations directory