migrations: add `migrate` subcommand to do database migrations
This allows us to migrate our database schema upwards with the migration
files embedded into the executable.

Adding a migration is done by creating a file containing SQL statements
to do the migration, and a filename of the format
"{version}_{name}.up.sql" where version is a four-number version number,
and should be incremental to the last migration file and name is a
underscore-separate name of what the migration does. After adding a
migration file you should run `go generate` to create the Go file that
is embedded into the executable