website: implemented api{/ping,/can-request,/request}
ping and can-request return the correct "main" json as is returned by
PHP right now, but omits the "meta" key that is returned by it. As it's
quite unlikely anyone is using that key anyway, it might be okay to omit
it in the final version too.

Implemented TrackCtx, that can supply a Song through URL parameters