website: implement /api/search (half)
Also decided to deprecate /api/metadata and /api/song, the former
because it was never used on the old website and seems to not be
functional. And the latter because the key changes daily for that
endpoint and thus nothing can be expecting the old behavior after we
switch over to a new version.

Rewrote /api/request to be /request and now returns the same messages as
the old website when returning json