mariadb: implemented part of NewsStorage required for v0 api
changed UpdatedAt and DeletedAt fields to pointers, since they are only
present if something actually got deleted or updated.

changed NewsStorage API to take an ID instead of the full post for
Comments and Delete. This since it doesn't really need any of the other
info present in the full post and just causes extra burden on the
caller.