website/admin: Changed session management to be a more concrete type
New session data should be added to the new SessionData struct in the
radio package.

Motivation to do this method is that in the end, having a proper type
for the session data should be more useful in the long run, and more
interchangeable if we stumble on a better session package.

PtrCodec is a thing because the Session struct would end up being
encoded only for a few steps in the session package and then get
passed back to us to decode again. This way we avoid the
encoding/decoding step