storage: remove "instance reusing"
This was using a very lackluster way of determining if an instance was
appropriate to return and resulted in a real instance being returned
to a test. Instead of a test instance that should've been freshly
created.

This feature was also already rarely used with most service entrypoints
making a singular storage instance they pass around at the init step.