website: refactor login middleware to follow the new code style
website: the shared input between all pages now lives in the middleware
package and is available to requests by retrieving it from the context.
This avoids a bunch of extra arguments everywhere and also avoids an import
cycle between shared and middleware due to the need of having access to the
LoginMiddleware data in the input creation.