rpc/manager: switched Manager.Status to Manager.CurrentStatus
which is a similar system to the other Manager endpoints. It's a stream
of status updates or if you just need one use `radio.OneOff` instead.

Quite a bit of the code using this can be optimized to actually use the
streaming properties, but for now they've all just been converted to use
OneOff instead.