radio: move OneOff to the util package
OneOff now has a bigger brother named StreamValue that keeps a
stream open in the background and continuously updates the value
until the context used to create it is canceled.

StreamValue also supports passing it a callback that is to be called
everytime a value arrives.

ircbot: move to the StreamValue for status updates

ircbot/AnnounceSong: now checks if a song is equal to the previous one
	instead of leaving this up to the caller.