manager: make the User optional in the direct User stream.
The old behaviour is kept for Status while if listening to the direct
stream for user changes you can now get a nil value indicating there
is nobody streaming. For Status.User it will keep using the previous
value that was there until a new User is passed to UpdateUser.