icecast: implement the icecast Source connection without net/http
net/http seems to really want to read the whole response if you use
the functions provided to read a response from an arbitrary reader.
But because we don't actually send any body from the proxy it just
hangs instead. This is probably a fault in the proxy but haven't
figured out how to fix that yet.