proxy: make the proxy more functional
Because our new icecast source client implementation for the streamer
and the one the proxy uses too uses net/http it uses http/1.1 and that
means we deal with Transfer-Encoding: chunked in some cases. This ment
that the old way we "accepted" the source client connection in the proxy
was broken because it didn't send a final chunk with chunked encoding.

This caused the new source client to hang on Body.Close until a timeout
was hit or the server hung up on us.