proxy: implement the function fields as methods instead
util: TypedValue now uses an atomic.Pointer internally so it lets you
set the value to nil. This does go at the cost of not having CompareAndSwap
since the comparison for that is considerable more complex now and I don't
feel like implementing it if we don't have a use for it.