proxy: add tests and redo mount cleanup
due to how we want to not immediately close the connection to the icecast
server when the last source on a mount leaves, the cleanup logic is a bit
convoluted. mounts are now cleaned up after mountTimeout has passed and no
new sources have appeared.

streamer/icecast: Metadata now passes options along