website/api/php: add PHP compatibility middleware
This is required to (probably) support the old android apps we have
hanging around. They use an obscure laravel feature where you can send
form data as a JSON object. The apps use this to send the CSRF token to
the request api, the middleware added should hopefully move this JSON token
to the appropiate header and be validated through standard gorille/csrf middleware