website/public: implement fix for the request url leaking
After requesting the url used for requesting '/v1/request' was leaking
into the newly rendered search page because it uses the request URL as is.
We patch it to instead get rewritten to be a '/search' url in input creation

website/public: implement test for csrfLegacyFix

The TestCSRFLegacyFix test should be doing about the same logic as the functions
used in the android app codebases.

mocks: generate extra mock types