radio, rpc, config, mocks: add AnnouncerService.AnnounceUser
ircbot: implement the new AnnounceUser method that should be
called when the active stream user changes, this is done internally
with a StreamValue.

AnnounceUser will announce the user on the main irc channel with similar
output as what the .dj command normally returns with the new user info.
It will also try to change the topic after a small delay such that spam
updates won't cause a slew of topic updates.

StreamerUserInfo now only returns the current dj instead of also allowing
you to set it.