util/eventstream: make the subs field private to the run goroutine
While not strictly required, tests ended up touching the subs field
and introducing race conditions that way. Moving it into the goroutine
makes it impossible to misuse it.

Added a LENGTH cmd to get the length of the above mentioned subs slice
so that tests can now do this without a race condition