radio: add Status.StreamUser
this field indicates the user that is actually streaming, that is it
is the user of whomever is connected to the proxy and has the highest
priority. This means if a user drops intermittingly or disconnects this
field is instantly set to nil.

for persistent user info the old User field should be used instead, this
field is only updated when a user change actually occurs excluding "no user"
scenarios