website/api/v1: fix flaky sse test
So the SSE Client Library that we use has an arbitrary limit of each
line being no longer than bufio.MaxScanTokenSize (which is 64*1024)
and the error handling is dramatic enough that this error does not
ever show its face to the user of said library.

We fixed the test by pretty-printing our test json so that it is split
across many lines.