search/bleve: write a custom tokenizer and query creator
the tokenizer is in essence a basic whitespace tokenizer but also
splitting on known japanese boundaries by using kagome

the query creator is a simple match query but using the correct analyzer
this time around (the same as the index creation).

the analyzer has been changed to use the new tokenizer and not use shingles
anymore. the ngram filter has been changed to not include the original term
anymore. some specific handling for cjk has also been removed and instead
leans on the new tokenizer and ngram filter (with the removal of the shingles
this works reasonably well (so far))