proxy: add support for removing a source forcefully (?)
this change makes Mount.RemoveSource close the connection on the
source we're removing. This is so it can be called from
non-RunMountSourceClient goroutines to forcefully remove the source.

also made the liveness check use the getter method since it needs to
acquire the read lock of the source to read this field, which it wasn't
doing before