radio: alter DJ.Theme to be a string (ThemeName)
Our source of truth for theme availability is the filesystem so storing
the themes in a table in the database is error-prone. Instead we remove
the table and we'll let the filesystem be the source of truth for theme
selection and other related things.

radio.Theme still exists but will be a type for the templates package
to expose with data from a file on the filesystem. This part isn't
implemented in this commit.

storage: we now only store a theme_name in the djs table, the themes
	table is dropped completely and we remove the theme_id column
	from the djs table, replaced by the theme_name

website/api/php: this legacy API apparently exposes the DJ theme as
	the theme_id, but since we don't have those anymore this is
	now just always set to 0, highly doubt anyone would be using
	this field too, but if they are they'll complain