streamer: add tracks type
this type is intended to solve issue 159 later on, but this is just
the isolated type with it not yet used in the streamer code.

tracks is supposed to be a bridge between the encoder function and
the icecast function such that the encoder can keep running aslong
as we haven't hit our preloaded duration target, that is if a song
is below 60 seconds in length, a second song should be prepared
instead of the encoder blocking.

it also encapsulates the encoder state slightly better such that we
can do a warm-start when doing a stop/start cycle instead of throwing
away all our state. this still needs to be implemented