streamer: handle restoring encoder songs better
previously this code kinda just yolo'd it and assumed there would ever
only be one song in the encoder entries, it also assumed that there
would be no race between the icecast routine telling the encoder to go
ahead with the next track while we don't prep the queue for this.

this ment that it was possible for the encoder to start working on the song
that was in the restored list instead of a new song. resulting in the same
song playing twice after a graceful restart.

there was also an assumption that the restored songs were still in the queue
and we still wanted to play them. this new code checks with the queue before
using restored entries, this also fixes the above problem of having it play
the same song twice since it properly sets up the reserved index in the queue