proxy: recover from panics in EventHandler
previously any panic would just crash the whole process, this is obviously
not what we want since the whole point of this separate event system was to
keep it nicely separated from the actual business logic. We now recover and
log any panics that occur.

fixed eventProxyStatus.UpdateLive being called with a nil SourceClient