website/shared: add NavBar type
this is an attempt at making the navbar and routes defined in code and
then rendered by templates, rather than each template having to figure
out what should be in the navbar.

current implementation is less than ideal since router paths and the
navbar paths are separate in the codebase but obviously should be
both taking from the same definitions. The navbar definitions are
put next to the router functions for this purpose for now.