search: Improve tokenization for exact matches
Add a new field to the search index that contains an
aggregate of the other search fields, with no ngram
tokenization. This is used to perform exact matches,
which are then boosted in score.

This works better than the previous solution, which
did not work for terms smaller than the ngram size.

Refs #212