storage/mariadb: put named query testing closer to the query definition
This is an attempt at localizing the named query testing, rather than
having a huge list of them in named_test.go where we might forget to
add one, we use a dummy function right next to the query definition
to check if the named query is valid.

Any invalid queries are stored in a map and will fail the new TestNamedQueries
test. We could also log or fail completely if invalid queries have been detected.