manager: fix tunein test to be slightly less failure prone
This thing is still just a race condition away from failing, but there
is nothing we can do about that right now. The problem is our zerolog
logger panicing when the test ends but StreamValue hasn't exited yet
from the ctx cancel.

There is no way for us to currently know about the state of the goroutine
so this added sleep is the best we can do for now. Go 1.24 introduced a
new experimental synctest package that might help with this in the future
but we can't use it yet.