search: fix tokenization for exact matching
This changes the analyzer for exact matching to be
the same as for fuzzy matching for better consistency
between the two.

It also changes the tokenizer to first split on whitespace,
and then only use kagome on the individual resulting tokens
when necessary. Otherwise, it uses the whitespace-split token
and does some punctuation filtering to construct the tokens.

This fixes the issue daniooo has with "u&i" not being
.ra-requestable, and also the issue tatsu has with
"konoka kokonoha" not being .ra-requestable.